Fix spi slave axi plug
Fixed an error in the "axi_plug" module of the IP slave spi: the winding length (used to indicate how many read transactions are required from the outside, starting from a base address) is continuously decremented by 1, so if the initial value was equal to 0, it turned out to be 31 (5-bit width), so more than one read transaction started erroneously. That is something to fix anyway.